+++ src/handle_query_contract_ui.c
@@ -18,6 +18,7 @@ static bool set_send_ui(ethQueryContractUI_t *msg, paraswap_parameters_t *contex
         case MULTI_SWAP_V4:
         case MEGA_SWAP_V4:
         case DIRECT_UNI_V3_SWAP:
+        case SWAP_EXACT_AMOUNT_IN_PRO:
             strlcpy(msg->title, "Send", msg->titleLength);
             break;
         case BUY_ON_UNI_FORK:
@@ -26,6 +27,7 @@ static bool set_send_ui(ethQueryContractUI_t *msg, paraswap_parameters_t *contex
         case SIMPLE_BUY:
         case BUY_ON_UNI_V4:
         case BUY_ON_UNI_FORK_V4:
+        case SWAP_EXACT_AMOUNT_OUT_PRO:
             strlcpy(msg->title, "Send Max", msg->titleLength);
             break;
         default:
@@ -61,6 +63,7 @@ static bool set_receive_ui(ethQueryContractUI_t *msg, paraswap_parameters_t *con
         case MULTI_SWAP_V4:
         case MEGA_SWAP_V4:
         case DIRECT_UNI_V3_SWAP:
+        case SWAP_EXACT_AMOUNT_IN_PRO:
             strlcpy(msg->title, "Receive Min", msg->titleLength);
             break;
         case BUY_ON_UNI_FORK:
@@ -69,6 +72,7 @@ static bool set_receive_ui(ethQueryContractUI_t *msg, paraswap_parameters_t *con
         case SIMPLE_BUY:
         case BUY_ON_UNI_V4:
         case BUY_ON_UNI_FORK_V4:
+        case SWAP_EXACT_AMOUNT_OUT_PRO:
             strlcpy(msg->title, "Receive", msg->titleLength);
             break;
         default: